image/image_caption/requirements.txt
torch
torchvision
Pillow
